Convert get_fee_info for testability
