+++ orderbook/src/api/filter.rs
@@ -1,13 +1,15 @@
 use super::handler;
 use crate::{database::OrderFilter, storage::Storage};
 use anyhow::Result;
+use chrono::{DateTime, FixedOffset, Utc};
 use hex::{FromHex, FromHexError};
 use model::{
     h160_hexadecimal,
     order::{Order, OrderCreation, OrderUid},
+    u256_decimal,
 };
-use primitive_types::H160;
-use serde::Deserialize;
+use primitive_types::{H160, U256};
+use serde::{Deserialize, Serialize};
 use std::{convert::Infallible, str::FromStr, sync::Arc};
 use warp::{
     http::StatusCode,
@@ -142,11 +144,36 @@ impl FromStr for H160Wrapper {
     }
 }
 
-pub fn get_fee_info() -> impl Filter<Extract = (impl Reply,), Error = Rejection> + Clone {
+pub fn get_fee_info_request() -> impl Filter<Extract = (H160,), Error = Rejection> + Clone {
     warp::path!("tokens" / H160Wrapper / "fee")
         .and(warp::get())
         .map(|token: H160Wrapper| token.0)
-        .and_then(handler::get_fee_info)
+}
+
+/// Fee struct being returned on fee API requests
+#[derive(PartialEq, Clone, Copy, Debug, Serialize, Deserialize)]
+#[serde(rename_all = "camelCase")]
+pub struct FeeInfo {
+    pub expiration_date: DateTime<Utc>,
+    #[serde(with = "u256_decimal")]
+    pub minimal_fee: U256,
+    pub fee_ratio: u32,
+}
+
+pub fn get_fee_info_response() -> impl Reply {
+    const STANDARD_VALIDITY_FOR_FEE_IN_SEC: i32 = 3600;
+    let fee_info = FeeInfo {
+        expiration_date: chrono::offset::Utc::now()
+            + FixedOffset::east(STANDARD_VALIDITY_FOR_FEE_IN_SEC),
+        minimal_fee: U256::zero(),
+        fee_ratio: 0u32,
+    };
+    with_status(warp::reply::json(&fee_info), StatusCode::OK)
+}
+
+pub fn get_fee_info() -> impl Filter<Extract = (impl Reply,), Error = Rejection> + Clone {
+    get_fee_info_request()
+        .and_then(|_token| async move { Result::<_, Infallible>::Ok(get_fee_info_response()) })
 }
 
 #[cfg(test)]
@@ -235,20 +262,21 @@ pub mod test_util {
     }
 
     #[tokio::test]
-    async fn get_fee_info_() {
-        let filter = get_fee_info();
-        let sell_token = String::from("0x000000000000000000000000000000000000000a");
-        let path_string = format!("/tokens/{}/fee", sell_token);
-        let post = || async {
-            request()
-                .path(&path_string)
-                .method("GET")
-                .reply(&filter)
-                .await
-        };
-        let response = post().await;
-        let body: handler::FeeInfo = serde_json::from_slice(response.body()).unwrap();
+    async fn get_fee_info_request_ok() {
+        let filter = get_fee_info_request();
+        let token = String::from("0x0000000000000000000000000000000000000000");
+        let path_string = format!("/tokens/{}/fee", token);
+        let request = request().path(&path_string).method("GET");
+        let result = request.filter(&filter).await.unwrap();
+        assert_eq!(result, H160::zero());
+    }
+
+    #[tokio::test]
+    async fn get_fee_info_response_() {
+        let response = get_fee_info_response().into_response();
         assert_eq!(response.status(), StatusCode::OK);
+        let body = response_body(response).await;
+        let body: FeeInfo = serde_json::from_slice(body.as_slice()).unwrap();
         assert_eq!(body.minimal_fee, U256::zero());
         assert_eq!(body.fee_ratio, 0);
         assert!(body.expiration_date.gt(&chrono::offset::Utc::now()))
